Contract fixes and improvements

+++ contracts/helpers/ERC20.sol
@@ -113,14 +113,14 @@ contract StandardToken is ERC20, BasicToken {
 	 * the first transaction is mined)
 	 * From MonolithDAO Token.sol
 	 */
-	function increaseApproval (address _spender, uint _addedValue) public returns (bool success) {
+	function increaseApproval (address _spender, uint256 _addedValue) public returns (bool success) {
 		allowed[msg.sender][_spender] = safeAdd(allowed[msg.sender][_spender], _addedValue);
 		Approval(msg.sender, _spender, allowed[msg.sender][_spender]);
 		return true;
 	}
 
-	function decreaseApproval (address _spender, uint _subtractedValue) public returns (bool success) {
-		uint oldValue = allowed[msg.sender][_spender];
+	function decreaseApproval (address _spender, uint256 _subtractedValue) public returns (bool success) {
+		uint256 oldValue = allowed[msg.sender][_spender];
 		if (_subtractedValue > oldValue) {
 			allowed[msg.sender][_spender] = 0;
 		} else {
