+++ test/TestBankeraToken.sol
@@ -9,11 +9,11 @@ contract TestBankeraToken {
   function testInitialBalanceUsingDeployedContract() public {
 	  BankeraToken meta = BankeraToken(DeployedAddresses.BankeraToken());
 
-    uint expected = 0;
-    uint256 totalSupply = 2500000000000000000;
+	  uint256 expected = 0;
+	  uint256 totalSupply = 2500000000000000000;
 
-    Assert.equal(meta.balanceOf(tx.origin), expected, "Owner should have 0 BankeraToken initially");
-    Assert.equal(meta.totalSupply(), totalSupply, "Incorrect totalSupply");
+	  Assert.equal(meta.balanceOf(tx.origin), expected, "Owner should have 0 BankeraToken initially");
+	  Assert.equal(meta.totalSupply(), totalSupply, "Incorrect totalSupply");
   }
